Fix macos build

+++ src/Lua/ComponentLuaFactory.h
@@ -112,14 +112,14 @@ class ComponentLuaFactory {
     const int height = SafeGetOptionalValue<int>(data, "height", 1);
     const glm::vec2 offset = SafeGetVec2(data, "offset");
     const auto collisionMask = SafeGetOptionalValue<int>(data, "collision_mask", Constants::kDefaultEntityMask);
-    const auto collisionMaskBits = EntityMask(collisionMask);
+    const auto collisionMaskBits = EntityMask(static_cast<unsigned long long>(collisionMask));
     return BoxColliderComponent(width, height, offset, collisionMaskBits);
   }
 
   static EntityMaskComponent CreateEntityMaskComponent(const sol::table& data) {
     using namespace LuaComponentHelpers;
     const auto raw = SafeGetOptionalValue<int>(data, "value", Constants::kDefaultEntityMask);
-    return EntityMaskComponent(EntityMask(raw));
+    return EntityMaskComponent(EntityMask(static_cast<unsigned long long>(raw)));
   }
 
   static HealthComponent CreateHealthComponent(const sol::table& data) {
@@ -135,9 +135,9 @@ class ComponentLuaFactory {
     const auto projectileDuration = SafeGetOptionalValue<float>(data, "projectile_duration", 1.0f);
     const int projectileDamage = SafeGetOptionalValue<int>(data, "hit_damage", 10);
     const auto collisionMask = SafeGetOptionalValue<int>(data, "collision_mask", Constants::kDefaultEntityMask);
-    const auto collisionMaskBits = EntityMask(collisionMask);
+    const auto collisionMaskBits = EntityMask(static_cast<unsigned long long>(collisionMask));
     const auto projectileMask = SafeGetOptionalValue<int>(data, "projectile_mask", 0);
-    const auto projectileMaskBits = EntityMask(projectileMask);
+    const auto projectileMaskBits = EntityMask(static_cast<unsigned long long>(projectileMask));
 
     auto component = ProjectileEmitterComponent(projectileVelocity, projectileDuration, repeatFrequency,
                                                 projectileDamage, collisionMaskBits, projectileMaskBits);
